chore: explode combined YAML to deref aliases and remove anchors
The `ireduce` yq function can "rearrange" the order of the top-level keys.
This becomes a problem for YAML files that have aliases declared AFTER it is referenced (anchor).

Specifically, the CircleCI CLI's `circleci config validate` throws an error when reading YAML
where an alias is declared AFTER the reference.
This is despite the YAML being valid.

Using the `explode` function helps to dereference aliases and remove the anchors.
This will produce YAML that the `circleci config validate` command can parse.